Improved XBRLInfo functions and Combining all data under singel data class
This job is addressing #6 #7 #8.
The functions in XBRLInfo function is being edited. It looks like the cal_file has good information to show the parent of the tag so now taking the parent from the cal_file. Weight and Order is also taken from the cal_file. pre_file is responsible for getting all the tags we will be dealing with and having it in a proper order.

I am also attempting to combine all the different dataclasses into a single XBRLNode class since having too many different ones makes it harder to follow.

Also, made some code refacoring with renaming some variable and organizing the functions in a more consise and understandable way.