Initial minimal editor implementation
The editor currently does not have any backend integration, however, all
the groundwork is ready. The current approach uses a simple, home-built
tokenizer that we can use to analyze user input and restructure
accordingly. The output of the tokenizer is currently used to produce  a
modified javascript code snippet that can be evaluated client-side to
retrieve necessary values. All intermediate retrieved and calculated
values are stored and shown to the user.