Improved xbrl_parse to get the right parent
There was an issue with C where it did not get the right parent since for one or two tags, there was no parent in the cal file within the balance sheet URI.
This is addressing #13 where we wanted to improve and get the right tag.

Now, we are going through the fs_URI and get all the parent. If it is still None after that, we will go through all URI to see if any other tables contains the parent of it.
We will assign it if we find it. Worst case, we will go through pre file again and get that parent, although that is not favorable.