Adding Condense Tags and Store file
Adding condense tags where we have our own maps to the us-gaap and company specifc tags. Ideally, we can have the user to make their own too but the users can make theirs too.
Also added a store file where we can store the processed data of the financial statements. This is still simple and we will need to work on to add complexity and scalability.