Adding the order of tags appearing in pre file and changing child to children
Added a new field in XBRLNode called lineup. Lineup will add the order in which the tag appears in the pre file.
It looks like pre file has the order of the tags appearing documented accurately so we can use this to keep the lineup of all tags.
There are some tags that are not in pre file and often times, these are Abtract tags that we don't neccessarily need so it should be ok.

Also renamed the field child to children to have a more accurate description since there can be more than one child per tag.