Started using Swagger for API Documentation
We will be using Swagger for API documentation to have code and related
comments describing the API within the sample file. Swagger UI also
allows us to quickly check/test the API to see if the responses are
behaving correctly. This will make it far easier to develop frontend
functionality when we reach that point. This work is related to #15