Storing multiple XBRLNode val and text
When storing the val and text field for XBRLNode, we will now store it as a list. This is due to the fact that some companies such as AAPL and C uses the same tags multiple times. We do this so we are able to store
all of the values and text that comes with each tag.

New format of XBRLNode.val:
    List[float]

New format of XBRLNode.text:
    List[str]

This addresses #20