Only add info to fs_fields for tags that exisit in assign_HTMLFact_to_XBRLNode
There was a bug where BAC bs from 2021 was failin to parse. This was due to the fact that we we trying to add info to fs_fields for a tag that did not exisit.
As a result, when doing fs_fields[tag], it overrode or bailed when we were tyring to add info to a key that did no exist. This was fixed by adding a check to make sure that the tag is in fs_field first.

This fixes the issue in #20. We can close #20.