Fixed bug regarding user cookies
To get around the issue of needing a non-HttpOnly cookie to change UI
views and checking with javascript to see if the user is still logged
in, we now also send a dummy cookie with the same lifetime parameters as
the actual authentication token. I've also started actually sketching
out the proper website and improving appearance. We will now be using
Chart.js for web plotting as the performance seems better and the UI is
simpler