Attempting to fix Unity issue for custom tag that is not ticker
U had a problem where its custom tag was not its ticker. Instead of U, it was Unity. I commented the lines of code that was causing it to parse incorrectly.

I also added a new block of code to try to find the tag another way if we cannot find the tag in our usual way. This way can work, since as far as I understand, the way they present these tags are usually just the company tag followed by an underscore, and then the full tag, or they have loc or numbers at the end. So I am looking through to see if the current split[i] has a number and if it does, I assume that it is not the tag. Since I have yet to see a tag that has numbers. It will probably be the case there is but I will pray that it does not reach this part of the code if it does. This will only be for companies with inconsistnet tag and ticker. This addresses #34.