Adding checks to see if EDGAR data is already downlaoded
This addresses #31.
We kept downloading the same data even if we already did. Added a small check in save_data_from_index to check if the directory already exist and if it does, we will not download it anymore.
It might be worth seeing if there are new files instead of just checking the folder already exist or not. I commented in the issue and will discuss it with Cale.