Bug on code for get_tag
In the get_tag function, there was a part of the code where we try to parse a tag. We had a code where we were referencing [i] but no object.
I added split in front to make it split[i]. Now, it properly gets the ith element of split. Previously, it was not working because it was doing [i] only.

This addresses #42 but TAYD is still not parsing properly. Will continue to work on the issue.