Refactoring of edgar_retrieve and adding new excel.py
The format of EDGAR seemed to change for retreiving the information so changed it. For older submissions, it used to store it in https://data.sec.gov/submissions/CIK{CIK}-submissions-001.json but now the submissions-001 is removed. Also did some minor refactorings.

I have also added a new excel.py for testing purposes only where we can use it to see if we are retrieving the balance sheet properly and will also retrieve the facts json file and pull out the tags that I specify. This currently has a lot of manual stuff so I need to add more autonomy to it in the future.