Adding cash flow and income statement spport for excel
excel.py has a cash flow and income statement support now. By default, I can get the historical data that is iXBRL and retrieve the 10-K IS and CF data.
We look at the duplicate tags to determine whether the tags are the same or not. Maybe also add support for text since some tags uses another tag but the text stays the same.